style(import-script): format code
